+++ src/components/GameDetailPage/style.js
@@ -1,0 +1,16 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles(theme => ({
+  card: {
+    height: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+  },
+  cardContent: {
+    flexGrow: 1
+  },
+  cardAction: {
+    display: 'flex',
+    justifyContent: 'center'
+  }
+}));
