+++ src/components/Gamelist/layout.js
@@ -1,23 +1,8 @@
 import React from 'react';
 import { Card, CardContent, CssBaseline, Grid, Typography, Container } from '@material-ui/core';
-import { makeStyles } from '@material-ui/core/styles';
 import { Link } from 'react-router-dom'
-import assignPlayers from '../assignPlayer'
-
-const useStyles = makeStyles(theme => ({
-  cardGrid: {
-    paddingTop: theme.spacing(8),
-    paddingBottom: theme.spacing(8),
-  },
-  card: {
-    height: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  cardContent: {
-    flexGrow: 1,
-  }
-}));
+import assignPlayers from '../../assignPlayer'
+import { useStyles } from './style'
 
 export default function Gamelist(props) {
   const classes = useStyles();
