+++ src/components/Homepage.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux'
-import GamelistContainer from './GamelistContainer'
+import Gamelist from './Gamelist'
 import { createGame } from '../actions'
 import CreateGameButton from './CreateGameButton'
 
@@ -13,7 +13,7 @@ class Homepage extends Component {
       <div>
         <div>
           <h1>These are the games you can join</h1>
-          <GamelistContainer />
+          <Gamelist />
         </div>
         <div className="creategame">
           {this.props.user.username && <CreateGameButton onClick={this.onClick} />}
